HSEARCH-4520 Introduce interface BooleanPredicateOptionsCollector
